requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Session==0.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
msgspec==0.18.6
openai==1.38.0
pip==24.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.29.0
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzlocal==5.2
urllib3==1.26.19
Werkzeug==3.0.3
python-dotenv==1.0.1
scipy==1.14.0
numpy==2.0.1